Create html form for adding criterion weights
Use javascript to generate a form for users to add weights. Form
dynamically creates as many labels and inputs as the number of criteria
selected. Weights can be input either by slider or number text box. Both
are scaled between 0 and 1 and automatically update the other when one
is changed.